+++ apps/api-journeys/src/modules/index.ts
@@ -8,5 +8,6 @@ export { journeyModule } from './journey'
 export { radioQuestionModule } from './radioQuestion'
 export { responseModule } from './response'
 export { signUpModule } from './signUp'
+export { stepModule } from './step'
 export { typographyModule } from './typography'
 export { videoModule } from './video'
